Split out SplitPostcode methods into concern.
